Move test into existing file to avoid export of shared classes.
