+++ packages/context/src/test/context-request_test.ts
@@ -10,13 +10,14 @@ import {property} from 'lit/decorators/property.js';
 import {
   ContextConsumer,
   ContextProvider,
+  Context,
   createContext,
   consume,
 } from '@lit/context';
 import {assert} from '@esm-bundle/chai';
 import {stripExpressionComments} from '@lit-labs/testing';
 
-export const simpleContext = createContext<number>('simple-context');
+const simpleContext = createContext<number>('simple-context');
 
 class SimpleContextProvider extends LitElement {
   private provider = new ContextProvider(this, {
@@ -29,7 +30,7 @@ class SimpleContextProvider extends LitElement {
   }
 }
 
-export class SimpleContextConsumer extends LitElement {
+class SimpleContextConsumer extends LitElement {
   // a one-time property fulfilled by context
   @consume({context: simpleContext})
   @property({type: Number})
@@ -73,6 +74,62 @@ suite('context-provider', () => {
     ) as SimpleContextProvider;
     assert.isDefined(provider);
     consumer = provider.querySelector(
+      'simple-context-consumer'
+    ) as SimpleContextConsumer;
+    assert.isDefined(consumer);
+  });
+
+  test(`consumer receives a context`, async () => {
+    assert.strictEqual(consumer.onceValue, 1000);
+    assert.strictEqual(consumer.subscribedValue, 1000);
+    assert.strictEqual(consumer.controllerContext.value, 1000);
+    await consumer.updateComplete;
+    assert.equal(
+      stripExpressionComments(consumer.shadowRoot!.innerHTML),
+      '1000'
+    );
+  });
+
+  test(`consumer receives updated context on provider change`, async () => {
+    assert.strictEqual(consumer.onceValue, 1000);
+    assert.strictEqual(consumer.subscribedValue, 1000);
+    assert.strictEqual(consumer.controllerContext.value, 1000);
+    await consumer.updateComplete;
+    assert.equal(
+      stripExpressionComments(consumer.shadowRoot!.innerHTML),
+      '1000'
+    );
+    provider.setValue(500);
+    assert.strictEqual(consumer.onceValue, 1000); // once value shouldn't change
+    assert.strictEqual(consumer.subscribedValue, 500);
+    assert.strictEqual(consumer.controllerContext.value, 500);
+    await consumer.updateComplete;
+    assert.equal(
+      stripExpressionComments(consumer.shadowRoot!.innerHTML),
+      '500'
+    );
+  });
+});
+
+suite('htmlelement-context-provider', () => {
+  let provider: ContextProvider<Context<unknown, number>, HTMLElement>;
+  let consumer: SimpleContextConsumer;
+
+  setup(async () => {
+    const container = document.createElement('div');
+    container.innerHTML = `
+       <simple-context-consumer></simple-context-consumer>
+     `;
+
+    provider = new ContextProvider(container, {
+      context: simpleContext,
+      initialValue: 2000,
+    });
+
+    document.body.appendChild(container);
+    provider.hostConnected();
+
+    consumer = container.querySelector(
       'simple-context-consumer'
     ) as SimpleContextConsumer;
     assert.isDefined(consumer);
